stash

+++ types/utils/helpers/arrays.d.ts
@@ -1,4 +1,4 @@
-import { ArrayType } from '../../utils/enums';
-declare const createTypedArray: (type: ArrayType, len: number) => number[] | Uint8ClampedArray<ArrayBuffer> | Int16Array<ArrayBuffer> | Float32Array<ArrayBuffer>;
+import { ArrayType } from '@/utils/enums';
+declare const createTypedArray: (type: ArrayType, len: number) => number[] | Float32Array<ArrayBuffer> | Int16Array<ArrayBuffer> | Uint8ClampedArray<ArrayBuffer>;
 declare function createSizedArray<T = unknown>(length: number): T[];
 export { createSizedArray, createTypedArray };
